move ParseRPCSerialPortSettings method to separate files
