+++ src/rpc/rpc_port_driver_list.cpp
@@ -1,4 +1,5 @@
 #include "rpc_port_driver_list.h"
+#include "rpc_serial_port_settings.h"
 #include "serial_port.h"
 #include "tcp_port.h"
 
@@ -87,16 +88,4 @@ void TRPCPortDriverList::ProcessException(const TRPCException& e, WBMQTT::TMqttR
             break;
         }
     }
-}
-
-TSerialPortConnectionSettings ParseRPCSerialPortSettings(const Json::Value& request)
-{
-    TSerialPortConnectionSettings res;
-    WBMQTT::JSON::Get(request, "baud_rate", res.BaudRate);
-    if (request.isMember("parity")) {
-        res.Parity = request["parity"].asCString()[0];
-    }
-    WBMQTT::JSON::Get(request, "data_bits", res.DataBits);
-    WBMQTT::JSON::Get(request, "stop_bits", res.StopBits);
-    return res;
 }
