+++ src/rpc/rpc_serial_port_settings.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "serial_port_settings.h"
+#include <wblib/json_utils.h>
+
+TSerialPortConnectionSettings ParseRPCSerialPortSettings(const Json::Value& request);
